jason - refactor CF_<domain objects> to utilize the Connection.cs class.  Much much cleaner.
git-svn-id: https://10.1.100.75:8443/svn/cloudfiles/trunk@42 a72f173c-07d4-0d44-999c-8bf02c4df871

+++ com.mosso.cloudfiles.integration.tests/Domain/CF/AccountSpecs.cs
@@ -1,4 +1,3 @@
-using System;
 using com.mosso.cloudfiles.domain;
 using com.mosso.cloudfiles.exceptions;
 using NUnit.Framework;
@@ -10,16 +9,16 @@ namespace com.mosso.cloudfiles.integration.tests.Domain.CF.AccountSpecs
     public class AccountIntegrationTestBase
     {
         protected UserCredentials userCredentials;
-        protected IAuthentication authentication;
+        protected IConnection connection;
         protected IAccount account;
 
         [SetUp]
         public void SetUp()
         {
             userCredentials = new UserCredentials(Constants.MOSSO_USERNAME, Constants.MOSSO_API_KEY);
-            authentication = new CF_Authentication(userCredentials);
+            connection = new Connection(userCredentials);
 
-            account = authentication.Authenticate();
+            account = connection.Account;
         }
     }
 
@@ -29,46 +28,39 @@ namespace com.mosso.cloudfiles.integration.tests.Domain.CF.AccountSpecs
         [Test]
         public void should_give_you_container_instance_when_successful()
         {
-            var containerName = Guid.NewGuid().ToString();
+
             var originalContainerCount = account.ContainerCount;
             var originalBytesUsed = account.BytesUsed;
             try
             {
-                account.CreateContainer(containerName);
-                Assert.That(account.ContainerExists(containerName), Is.True);
+                account.CreateContainer(Constants.CONTAINER_NAME);
+                Assert.That(account.ContainerExists(Constants.CONTAINER_NAME), Is.True);
                 Assert.That(account.ContainerCount, Is.EqualTo(originalContainerCount + 1));
                 Assert.That(account.BytesUsed, Is.EqualTo(originalBytesUsed + 0));
             }
             finally
             {
-                account.DeleteContainer(containerName);
-                Assert.That(account.ContainerExists(containerName), Is.False);
+                if(account.ContainerExists(Constants.CONTAINER_NAME))
+                    account.DeleteContainer(Constants.CONTAINER_NAME);
+                Assert.That(account.ContainerExists(Constants.CONTAINER_NAME), Is.False);
                 Assert.That(account.ContainerCount, Is.EqualTo(originalContainerCount));
                 Assert.That(account.BytesUsed, Is.EqualTo(originalBytesUsed));
             }
-            
-
-            Assert.That(account.StorageUrl.ToString().Contains("https://storage.clouddrive.com/v1/MossoCloudFS_"), Is.True);
-            Assert.That(account.StorageToken.Length, Is.EqualTo(36));
-            Assert.That(account.AuthToken.Length, Is.EqualTo(36));
-            Assert.That(account.CDNManagementUrl.ToString().Contains("https://cdn.clouddrive.com/v1/MossoCloudFS_"), Is.True);
-
         }
 
         [Test]
         [ExpectedException(typeof(ContainerAlreadyExistsException))]
         public void should_throw_container_already_exists_exception_if_the_container_already_exists()
         {
-            var containerName = Guid.NewGuid().ToString();
             try
             {
-                account.CreateContainer(containerName);
-                account.CreateContainer(containerName);
+                account.CreateContainer(Constants.CONTAINER_NAME);
+                account.CreateContainer(Constants.CONTAINER_NAME);
             }
             finally
             {
-                account.DeleteContainer(containerName);
-                Assert.That(account.ContainerExists(containerName), Is.False);
+                account.DeleteContainer(Constants.CONTAINER_NAME);
+                Assert.That(account.ContainerExists(Constants.CONTAINER_NAME), Is.False);
             }
         }
     }
@@ -80,30 +72,33 @@ namespace com.mosso.cloudfiles.integration.tests.Domain.CF.AccountSpecs
         [ExpectedException(typeof(ContainerNotEmptyException))]
         public void should_throw_container_not_empty_exception_if_the_container_not_empty()
         {
-            var containerName = Guid.NewGuid().ToString();
             IContainer container = null;
             var originalContainerCount = account.ContainerCount;
             var originalBytesUsed = account.BytesUsed;
             try
             {
-                
-                container = account.CreateContainer(containerName);
+
+                container = account.CreateContainer(Constants.CONTAINER_NAME);
                 container.AddObject(Constants.StorageItemName);
                 Assert.That(container.ObjectExists(Constants.StorageItemName), Is.True);
                 Assert.That(account.ContainerCount, Is.EqualTo(originalContainerCount + 1));
                 Assert.That(account.BytesUsed, Is.EqualTo(originalBytesUsed + 34));
 
-                account.DeleteContainer(containerName);
+                account.DeleteContainer(Constants.CONTAINER_NAME);
             }
             finally
             {
                 Assert.That(account.ContainerCount, Is.EqualTo(originalContainerCount + 1));
                 Assert.That(account.BytesUsed, Is.EqualTo(originalBytesUsed + 34));
 
-                container.DeleteObject(Constants.StorageItemName);
-                Assert.That(container.ObjectExists(Constants.StorageItemName), Is.False);
-                account.DeleteContainer(containerName);
-                Assert.That(account.ContainerExists(containerName), Is.False);
+                if(container != null && account.ContainerExists(Constants.CONTAINER_NAME))
+                {
+                    container.DeleteObject(Constants.StorageItemName);
+                    Assert.That(container.ObjectExists(Constants.StorageItemName), Is.False);
+                    account.DeleteContainer(Constants.CONTAINER_NAME);    
+                }
+                
+                Assert.That(account.ContainerExists(Constants.CONTAINER_NAME), Is.False);
                 Assert.That(account.ContainerCount, Is.EqualTo(originalContainerCount));
                 Assert.That(account.BytesUsed, Is.EqualTo(originalBytesUsed));
             }
@@ -116,13 +111,12 @@ namespace com.mosso.cloudfiles.integration.tests.Domain.CF.AccountSpecs
         [Test]
         public void should_keep_count_of_each_container_and_object()
         {
-            var containerName = Guid.NewGuid().ToString();
             IContainer container = null;
             var originalContainerCount = account.ContainerCount;
             try
             {
-                
-                container = account.CreateContainer(containerName);
+
+                container = account.CreateContainer(Constants.CONTAINER_NAME);
                 container.AddObject(Constants.StorageItemName);
                 container.AddObject(Constants.HeadStorageItemName);
                 Assert.That(container.ObjectExists(Constants.StorageItemName), Is.True);
@@ -131,13 +125,17 @@ namespace com.mosso.cloudfiles.integration.tests.Domain.CF.AccountSpecs
             }
             finally
             {
-                container.DeleteObject(Constants.StorageItemName);
-                container.DeleteObject(Constants.HeadStorageItemName);
-                Assert.That(container.ObjectCount, Is.EqualTo(0));
-                Assert.That(container.ObjectExists(Constants.StorageItemName), Is.False);
-                Assert.That(container.ObjectExists(Constants.HeadStorageItemName), Is.False);
-                account.DeleteContainer(containerName);
-                Assert.That(account.ContainerExists(containerName), Is.False);
+                if (container != null && account.ContainerExists(Constants.CONTAINER_NAME))
+                {
+                    container.DeleteObject(Constants.StorageItemName);
+                    container.DeleteObject(Constants.HeadStorageItemName);
+                    Assert.That(container.ObjectCount, Is.EqualTo(0));
+                    Assert.That(container.ObjectExists(Constants.StorageItemName), Is.False);
+                    Assert.That(container.ObjectExists(Constants.HeadStorageItemName), Is.False);
+                    account.DeleteContainer(Constants.CONTAINER_NAME);
+                }
+
+                Assert.That(account.ContainerExists(Constants.CONTAINER_NAME), Is.False);
                 Assert.That(account.ContainerCount, Is.EqualTo(originalContainerCount));
             }
         }
@@ -149,18 +147,17 @@ namespace com.mosso.cloudfiles.integration.tests.Domain.CF.AccountSpecs
         [Test]
         public void should_return_json_string_with_container_names_and_item_count_and_bytes_used()
         {
-            var containerName = Guid.NewGuid().ToString();
             try
             {
-                account.CreateContainer(containerName);
-                var expectedJson = "[{\"name\": \"" + containerName + "\", \"count\": 0, \"bytes\": 0}]";
+                account.CreateContainer(Constants.CONTAINER_NAME);
+                var expectedJson = "[{\"name\": \"" + Constants.CONTAINER_NAME + "\", \"count\": 0, \"bytes\": 0}]";
 
                 Assert.That(account.JSON, Is.EqualTo(expectedJson));
             }
             finally
             {
-                account.DeleteContainer(containerName);
-                Assert.That(account.ContainerExists(containerName), Is.False);
+                account.DeleteContainer(Constants.CONTAINER_NAME);
+                Assert.That(account.ContainerExists(Constants.CONTAINER_NAME), Is.False);
             }
             
         }
@@ -184,18 +181,17 @@ namespace com.mosso.cloudfiles.integration.tests.Domain.CF.AccountSpecs
         [Test]
         public void should_return_xml_document_with_container_names_and_item_count_and_bytes_used()
         {
-            var containerName = Guid.NewGuid().ToString();
             try
             {
-                account.CreateContainer(containerName);
-                var expectedXml = "<?xml version=\"1.0\" encoding=\"UTF-8\"?><account name=\"MossoCloudFS_5d8f3dca-7eb9-4453-aa79-2eea1b980353\"><container><name>" + containerName + "</name><count>0</count><bytes>0</bytes></container></account>";
+                account.CreateContainer(Constants.CONTAINER_NAME);
+                var expectedXml = "<?xml version=\"1.0\" encoding=\"UTF-8\"?><account name=\"MossoCloudFS_5d8f3dca-7eb9-4453-aa79-2eea1b980353\"><container><name>" + Constants.CONTAINER_NAME + "</name><count>0</count><bytes>0</bytes></container></account>";
 
                 Assert.That(account.XML.InnerXml, Is.EqualTo(expectedXml));
             }
             finally
             {
-                account.DeleteContainer(containerName);
-                Assert.That(account.ContainerExists(containerName), Is.False);
+                account.DeleteContainer(Constants.CONTAINER_NAME);
+                Assert.That(account.ContainerExists(Constants.CONTAINER_NAME), Is.False);
             }
         }
     }
